cleanup and refactor. tests pending
